Refactored a bit in order to avoid clutter on App.jsx and set a better standard for separation of concerns.
