+++ client/src/App.jsx
@@ -27,7 +27,6 @@ export default class App extends React.Component {
         this.onChangeSighUp = this.onChangeSignUp.bind(this);
         this.onClickSignUp = this.onClickSignUp.bind(this);
         this.onClickLogin = this.onClickLogin.bind(this);
-        this.onAddToPantry = this.onAddToPantry.bind(this);
         this.onChangeAddItem = this.onChangeAddItem.bind(this);
     }
 
@@ -54,17 +53,6 @@ export default class App extends React.Component {
   }
 
 
-  onAddToPantry () {
-    const addItem = {
-      item: this.state.item_name,
-      exp: this.state.expiration
-    };
-    axios.post('/addingtopantry', addItem)
-      .then( response => {
-        console.log(response.data);
-      })
-  }
-
   onChangeAddItem (event) {
     this.setState ({
         [event.target.name]: event.target.value
@@ -81,8 +69,9 @@ export default class App extends React.Component {
       return (
         <div>
 
-          {/* <Welcome change={this.onChangeSignUp} onClickLogin={this.onClickLogin} onClickSignUp={this.onClickSignUp} name={this.state.name} username={this.state.userName} password={this.state.password} email={this.state.email} SignUp={this.state.SignUp} Login={this.state.Login} /> */}
-          <Pantry onAddToPantry={this.onAddToPantry} expiration={this.state.expiration} item_name={this.state.item_name} onChangeAddItem={this.onChangeAddItem}/>
+          <Welcome change={this.onChangeSignUp} onClickLogin={this.onClickLogin} onClickSignUp={this.onClickSignUp} name={this.state.name} username={this.state.userName} password={this.state.password} email={this.state.email} SignUp={this.state.SignUp} Login={this.state.Login} />
+          <Pantry onAddToPantry={this.onAddToPantry} expiration={this.state.expiration} item_name={this.state.item_name} 
+          onChangeAddItem={this.onChangeAddItem} item_name={this.state.item_name} expiration={this.state.expiration}/>
         </div>
       );
     }
